% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sorting.R
\name{add_sorting}
\alias{add_sorting}
\title{Add sorting}
\usage{
add_sorting(x, method = "count_desc")
}
\arguments{
\item{x}{A categorical vector}

\item{method}{One of:
\itemize{
  \item{"count_desc"}{Descending count}
}}
}
\description{
Sort based on some criteria
}
