% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{histogram}
\alias{histogram}
\title{Histogram}
\usage{
histogram(x, color_num = 1, ...)
}
\arguments{
\item{x}{a numeric vector,}

\item{color_num}{An integer  between}

\item{...}{parameters passed to plotly::plot_ly()}
}
\description{
Creates a Plotly object histogram in a single step
}
\examples{
histogram(mtcars$mpg, nbinsx=10)

}
