% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_integer}
\alias{is_integer}
\title{Is integer}
\usage{
is_integer(x)
}
\arguments{
\item{x}{a numeric value or vector}
}
\description{
Check whether a numeric is an integer or not
}
\examples{
is_integer(2)
is_integer(c(2,2.5))

}
