% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_step_func.R
\name{m_step_regression}
\alias{m_step_regression}
\title{A wrapper function providing the quantities related to the M-step as outlined in citation}
\usage{
m_step_regression(Y_dat, W, W2, X, df = NULL, Int = TRUE)
}
\arguments{
\item{Y_dat}{A matrix containing the outcome \code{Y}}

\item{W}{Quantity \eqn{W} as outlined in citation, output from W_update_fun}

\item{W2}{Quantity \eqn{W^2} as outlined in citation, output from W_update_fun}

\item{X}{A matrix or dataframe of other predictors to account for}
}
\value{
A list including a vector of coefficients' estimates for calibration model, the model's residual sum of squares, the model's R-squared, predicted values for the outcome Y, a vector containing the diagonal of the 'hat' matrix,  model residuals, variance covariance matrix, and a dataframe containing t-test statistics of estimated coefficients and associated p-values and degrees of freedom
}
\description{
A wrapper function providing the quantities related to the M-step as outlined in citation
}
\examples{
#not run
#mod <- m_step_regression(Y, W_ast, W_ast_var + W_ast^2, X)
}
