% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtr_func.R
\name{mtr_func}
\alias{mtr_func}
\title{A function providing quantities related to the E-step after applying various corrections}
\usage{
mtr_func(E_step, alpha, signal = NULL)
}
\arguments{
\item{E_step}{The object returned by the E_step_func function}

\item{alpha}{Type I error; significance level}

\item{signal}{An optional argument, a vector of true non-null coefficients to calculate the true and false discovery rates}
}
\value{
A list of results adjusted via the FDR correction, the Benjamini-Hochberg-Yekutieli BY correction, and the Bonferroni correction
}
\description{
A function providing quantities related to the E-step after applying various corrections
}
\examples{
#not run
#MTR_res <- mtr_func(E_step, alpha, signal)
}
