% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{palette_feather}
\alias{palette_feather}
\title{Feather palette}
\format{
An object of class \code{character} of length 9.
}
\usage{
palette_feather
}
\description{
Inspired by 'You Had Your Soul With You' by The National
}
\keyword{datasets}
