% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_unhidem_func.R
\name{plot_unhidem_func}
\alias{plot_unhidem_func}
\title{A function to plot various unhidem algorithm results}
\usage{
plot_unhidem_func(full_res, test_plot, alpha)
}
\arguments{
\item{full_res}{The output from the unhidem function}

\item{test_plot}{A logical value (true/false) indicating whether to include test data when plotting}

\item{alpha}{Target false discovery rate; only used if test_plot = FALSE.}
}
\value{
Plots a figure but does not return an object
}
\description{
A function to plot various unhidem algorithm results
}
\examples{
#not run
#plot_unhidem_func(full_res, test_plot = !is.null(Y_test), alpha = alpha)
}
