% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_hprobe_func.R
\name{predict_hprobe_func}
\alias{predict_hprobe_func}
\title{Obtaining predictions, confidence intervals and prediction intervals from probe}
\usage{
predict_hprobe_func(res, X, V, Z = NULL, alpha = 0.05, X_2 = NULL)
}
\arguments{
\item{res}{The results from the probe function.}

\item{X}{A matrix containing the predictors on which to apply the probe algorithm}

\item{V}{A design matrix of predictors for the variance model.}

\item{Z}{(optional) A matrix or dataframe of predictors not subjected to the sparsity assumption to account for.}

\item{alpha}{significance level for (\eqn{100(1-\alpha)\%}) credible and prediction intervals.}

\item{X_2}{(optional) Square of X matrix.}
}
\value{
A dataframe with predictions, credible intervals, and prediction intervals for each new observation.
}
\description{
A function providing predictions, along with \eqn{(1-\alpha)*100\%} credible, and prediction intervals for new observations.
}
\examples{
### Example
#not run
# pred_res <- predict_probe_func(full_res, X = X_test, Z = NULL, alpha = alpha)
# head(pred_res)

}
\references{
McLain, A. C., Zgodic, A., & Bondell, H. (2022). Sparse high-dimensional linear regression with a partitioned empirical Bayes ECM algorithm. arXiv preprint arXiv:2209.08139.
}
