% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_unhidem_func.R
\name{predict_unhidem_func}
\alias{predict_unhidem_func}
\title{Obtaining predictions, confidence intervals and prediction intervals from UNHIDEM}
\usage{
predict_unhidem_func(full_res, Z = Z_test, alpha = alpha)
}
\arguments{
\item{full_res}{The object of the unhidem function}

\item{Z}{A matrix containing the predictors on which to apply the UNHIDEM algorithm}

\item{X}{A matrix or dataframe of predictors not subjected to the sparsity assumption to account for (if present).}

\item{alpha}{Type I error; significance level}
}
\value{
A dataframe with predictions, confidence intervals, and prediction intervals for each new observation.
}
\description{
A function providing predictions, along with (1-alpha)*100% confidence intervals, and prediction intervals for new observations
}
\examples{
#not run
#predict_unhidem_func(full_res, Z = Z_test, alpha = alpha)


}
