\name{unhidem-package}
\alias{unhidem-package}
\alias{unhidem}
\docType{package}
\title{\packageTitle{unhidem}}
\description{\packageDescription{unhidem}}
\section{Package Content}{\packageIndices{unhidem}}
\author{\packageAuthor{unhidem}}
\section{Maintainer}{\packageMaintainer{unhidem}}
\keyword{package}
