% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hprobe.R
\name{hprobe}
\alias{hprobe}
\title{Fitting PaRtitiOned empirical Bayes Ecm (PROBE) algorithm to sparse high-dimensional linear models with heterogeneous variance.}
\usage{
hprobe(Y, X, V, Z = NULL, ep = 0.1, maxit = 10000, Y_test = NULL, X_test = NULL, 
Z_test = NULL, V_test = NULL, verbose = FALSE, signal = NULL, eta_i = NULL, alpha = 0.05, 
plot_ind = FALSE, adj = 5)
}
\arguments{
\item{Y}{The outcome variable.}

\item{X}{An \code{n x M} matrix of sparse predictors variables.}

\item{V}{A design matrix of predictors for the variance model (including an intercept).}

\item{Z}{(optional) An \code{n x p} matrix or dataframe of other predictors not subjected to the sparsity assumption.}

\item{ep}{Value against which to compare convergence criterion (default = 0.1).}

\item{maxit}{Maximum number of iterations the algorithm will run for (default = 10000).}

\item{Y_test}{(optional) Test Y data used plotting purposes only (doesn't impact results).}

\item{X_test}{(optional) Test X data used plotting purposes only (doesn't impact results).}

\item{Z_test}{(optional) Test Z data used plotting purposes only (doesn't impact results).}

\item{V_test}{(optional) Test V data used plotting purposes only (doesn't impact results).}

\item{verbose}{A logical (true/false) value whether to print algorithm iteration progress and summary quantities (default = FALSE).}

\item{signal}{(optional) A vector of indicies of the true non-null coefficients. This is used to calculate the true and false discovery rates by iteration for simulated data. Used plotting purposes only (doesn't impact results).}

\item{eta_i}{(optional) A vector of the true signal. This is used to calculate the MSE by iteration for simulated data. Used plotting purposes only (doesn't impact results).}

\item{alpha}{(optional) significance level}

\item{plot_ind}{A logical values (True/False) for whether to output plots on algorithm results and progress (default = FALSE)}

\item{adj}{Bandwidth parameter for empirical Bayes E-step. The bandwidth will be equal to \code{adj} times Silverman's 'rule of thumb' (default = 2).}
}
\value{
A list including 

\code{beta_ast_hat} MAP estimates of the regression coefficients (\eqn{\beta^\ast}),

\code{beta_hat, beta_hat_var} MAP estimates of the posterior expectation (beta_hat) and variance (beta_hat_var) of the prior mean (\eqn{\beta}) of the regression coefficients assuming \eqn{\gamma=1}, 

\code{gamma_hat} the posterior expectation of the latent \eqn{\gamma} variables, 

\code{sigma2_est} MAP estimate of the residual variance, 

\code{E_step} full results of the final E_step, 

\code{Calb_mod} results of first (\eqn{\alpha_0}) part of the M-step,  

\code{count} the total number of iterations before convergence.
}
\description{
A wrapper function for the H-PROBE algorithm.
}
\examples{
### Example
data(h_Sim_data)
attach(h_sim_data)

# Run Analysis
res <- hprobe(Y = Y, X = X, V = V)
 
# Predicting for test data
pred_res <- predict_hprobe_func(res, X_test, V = V_test)
sqrt(mean((Y_test - pred_res$Pred)^2))
head(cbind(Y_test, pred_res))

plot(Y_test, pred_res$Pred, ylab = "Prediction", xlab = "Test Outcome")
abline(coef = c(0,1))

# Proportion of explained variance
1 - var(Y_test - pred_res$Pred)/var(Y_test)

## Omega coefficients (versus true values)
cbind(omega_tr, res$omega)

## True versus estimated beta coeffiecients
plot(beta_tr, 
     res$beta_ast_hat, 
     xlab = "True Beta", 
     ylab = "Estimated Beta")
abline(coef = c(0,1))

## Confusion matrix of true versus estimated signals using 0.5 cutoff.
table(beta_tr==0, res$gamma_hat<0.5)


}
\references{
McLain, AC, A Zgodic, H Bondell (2025). Sparse high-dimensional linear regression with a partitioned empirical Bayes ECM algorithm. **Computational Statistics and Data Analysis** 207, 108146.
Zgodic, A., Bai, R., Zhang, J., Wang, Y., Rorden, C., & McLain, A. (2023). Quantifying predictive uncertainty of aphasia severity in stroke patients with sparse heteroscedastic Bayesian high-dimensional regression. arXiv preprint arXiv:2309.08783.
}
\seealso{
predict_probe_func to obtain predictions, credible intervals and prediction intervals from PROBE.
}
