% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_probe_func.R
\name{predict_probe_func}
\alias{predict_probe_func}
\title{Obtaining predictions, confidence intervals and prediction intervals from probe}
\usage{
predict_probe_func(res, X, Z = NULL, alpha = 0.05, X_2 = NULL)
}
\arguments{
\item{res}{The results from the probe function.}

\item{X}{A matrix containing the predictors on which to apply the probe algorithm}

\item{Z}{(optional) A matrix or dataframe of predictors not subjected to the sparsity assumption to account for.}

\item{alpha}{significance level for (\eqn{100(1-\alpha)\%}) credible and prediction intervals.}

\item{X_2}{(optional) Square of X matrix.}
}
\value{
A dataframe with predictions, credible intervals, and prediction intervals for each new observation.
}
\description{
A function providing predictions, along with \eqn{(1-\alpha)*100\%} credible, and prediction intervals for new observations.
}
\examples{
### Example
#not run
# pred_res <- predict_probe_func(full_res, X = X_test, Z = NULL, alpha = alpha)
# head(pred_res)

}
\references{
\itemize{ \item McLain, AC, A Zgodic, H Bondell (2025). Sparse high-dimensional linear regression with a partitioned empirical Bayes ECM algorithm. \textit{Computational Statistics and Data Analysis} 207, 108146.
\item Zgodic, A., Bai, R., Zhang, J., Wang, Y., Rorden, C., & McLain, A. (2023). Quantifying predictive uncertainty of aphasia severity in stroke patients with sparse heteroscedastic Bayesian high-dimensional regression. arXiv preprint arXiv:2309.08783.}
}
