% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_probe_func.R
\name{predict_probe_func}
\alias{predict_probe_func}
\title{Obtaining predictions, confidence intervals and prediction intervals from probe}
\usage{
predict_probe_func(res, Z, X, alpha, Z_2)
}
\arguments{
\item{res}{The results from the probe function.}

\item{Z}{A matrix containing the predictors on which to apply the probe algorithm}

\item{X}{(optional) A matrix or dataframe of predictors not subjected to the sparsity assumption to account for.}

\item{alpha}{Significance level for \eqn{(1-\alpha)*100\%} credible, and prediction intervals.}

\item{Z_2}{(optional) Square of Z matrix.}
}
\value{
A dataframe with predictions, confidence intervals, and prediction intervals for each new observation.
}
\description{
A function providing predictions, along with \eqn{(1-\alpha)*100\%} credible, and prediction intervals for new observations.
}
\examples{
### Example
data(Sim_data)
attach(Sim_data)
alpha <- 0.05
plot_ind <- TRUE

# Run the analysis. Y_test and Z_test are included for plotting purposes only
full_res <- probe( Y = Y, Z = Z, alpha = alpha, plot_ind = plot_ind, Y_test = Y_test, Z_test = Z_test)

# Predicting for test data
pred_res <- predict_probe_func(full_res, Z = Z_test, X = NULL, alpha = alpha)
head(pred_res)

}
