% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probe-package.R
\docType{package}
\name{probe-package}
\alias{probe-package}
\alias{_PACKAGE}
\title{probe: Sparse high-dimensional linear regression with a PaRtitiOned empirical Bayes Ecm (PROBE) algorithm}
\description{
This package contains functions to fit an efficient and powerful Bayesian approach for sparse high-dimensional linear regression. Minimal prior assumptions on the parameters are used through the use of plug-in empirical Bayes estimates of hyperparameters. Efficient maximum a posteriori (MAP) estimation is completed through a Parameter-Expanded Expectation-Conditional-Maximization (PX-ECM) algorithm. The PX-ECM results in a robust computationally efficient coordinate-wise optimization, which adjusts for the impact of other predictor variables. The completion of the E-step uses an approach motivated by the popular two-groups approach to multiple testing. The result is a PaRtitiOned empirical Bayes Ecm (PROBE) algorithm applied to sparse high-dimensional linear regression, which can be completed using one-at-a-time or all-at-once type optimization.
}
\details{
Examples for applying PROBE to sparse high-dimensional linear regression are given
for one-at-a-time \code{\link{probe_one}} or all-at-once \code{\link{probe}} type optimization.
}
\references{
\itemize{ \item McLain, A. C., Zgodic, A., & Bondell, H. (2022). Sparse high-dimensional linear regression with a partitioned empirical Bayes ECM algorithm. arXiv preprint arXiv:2209.08139.}
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/alexmclain/PROBE/issues}
}

}
\author{
\strong{Maintainer}: Alexander McLain \email{mclaina@mailbox.sc.edu}

Authors:
\itemize{
  \item Anja Zodiac
}

}
